+++ application/services/custom_services/patientdetails/src/main/java/com/geppetto/MediRecords/service/serviceimpl/BillingdetailsServiceImpl.java
@@ -9,13 +9,12 @@ import java.util.List;
 import org.springframework.data.domain.PageRequest;
 import java.util.Map;
 import org.springframework.data.domain.Page;
-import com.geppetto.MediRecords.util.ConstructQuery;
+import com.geppetto.MediRecords.util.PatientDetailsUtil;
 import org.springframework.data.domain.Pageable;
 import com.geppetto.MediRecords.repository.BillingdetailsRepository;
 import java.util.stream.Collectors;
 import org.springframework.data.jpa.domain.Specification;
 
-import org.springframework.beans.BeanUtils;
 import org.springframework.stereotype.Service;
 import lombok.extern.slf4j.Slf4j;
 import lombok.RequiredArgsConstructor;
@@ -38,7 +37,7 @@ public class BillingdetailsServiceImpl implements BillingdetailsService {
     */
     private final BillingdetailsDao billingdetailsDao;
     private final BillingdetailsRepository billingdetailsRepository;
-    private final ConstructQuery constructQuery;
+    private final PatientDetailsUtil patientDetailsUtil;
 
   
     /**
@@ -49,13 +48,14 @@ public class BillingdetailsServiceImpl implements BillingdetailsService {
     */
     @Override
     public BillingdetailsDto  createBillingdetails(BillingdetailsDto billingdetailsDto) {
-    log.info("Entering createBillingdetails method");
-    Billingdetails billingdetails = new Billingdetails();
-    BeanUtils.copyProperties(billingdetailsDto, billingdetails);
-    Billingdetails createdBillingdetails = billingdetailsDao.createBillingdetails(billingdetails);
-    BeanUtils.copyProperties(createdBillingdetails, billingdetailsDto);
-    log.info("Exiting createBillingdetails method");
-    return billingdetailsDto;
+        log.info("Entering createBillingdetails method");
+    
+        Billingdetails billingdetails = patientDetailsUtil.toEntity(billingdetailsDto);
+        Billingdetails createdBillingdetails = billingdetailsDao.createBillingdetails(billingdetails);
+        billingdetailsDto = patientDetailsUtil.toDto(createdBillingdetails);
+        
+        log.info("Exiting createBillingdetails method");
+        return billingdetailsDto;
     }
   
     /**
@@ -67,18 +67,18 @@ public class BillingdetailsServiceImpl implements BillingdetailsService {
     */
     @Override
     public BillingdetailsDto  getBillingdetailsById(String id) {
-    log.info("Entering getBillingdetailsById method for ID: {}", id);
-    return billingdetailsDao.getBillingdetailsById(id)
-        .map(billingdetails -> {
-            BillingdetailsDto dto = new BillingdetailsDto();
-            BeanUtils.copyProperties(billingdetails, dto);
-            log.info("Exiting getBillingdetailsById method for ID: {}", id);
-            return dto;
-        })
-        .orElseThrow(() -> {
+        log.info("Entering getBillingdetailsById method for ID: {}", id);
+        
+        Billingdetails billingdetails = billingdetailsDao.getBillingdetailsById(id)
+            .orElseThrow(() -> {
             log.warn("No billingdetails found for ID: {}", id);
             return new EntityNotFoundException("Data not found for ID: " + id);
         });
+
+        BillingdetailsDto billingdetailsDto = patientDetailsUtil.toDto(billingdetails);
+        
+        log.info("Exiting getBillingdetailsById method for ID: {}", id);
+        return billingdetailsDto;
     }
   
     /**
@@ -88,16 +88,17 @@ public class BillingdetailsServiceImpl implements BillingdetailsService {
     */
     @Override
     public Page<BillingdetailsDto>  getAllBillingdetails(int page, int size) {
-    log.info("Entering getAllBillingdetails method");
-    Pageable pageable = PageRequest.of(page, size);
-    Page<Billingdetails> billingdetailsPage = billingdetailsDao.getAllBillingdetails(pageable);
-    Page<BillingdetailsDto> billingdetailsDtoPage = billingdetailsPage.map(billingdetails -> {
-        BillingdetailsDto dto = new BillingdetailsDto();
-        BeanUtils.copyProperties(billingdetails, dto);
-        return dto;
-    });
-    log.info("Exiting getAllBillingdetails method");
-    return billingdetailsDtoPage;
+        log.info("Entering getAllBillingdetails method");
+    
+        Pageable pageable = PageRequest.of(page, size);
+        Page<Billingdetails> billingdetailsPage = billingdetailsDao.getAllBillingdetails(pageable);
+        Page<BillingdetailsDto> billingdetailsDtoPage = billingdetailsPage.map(billingdetails -> {
+            BillingdetailsDto dto = patientDetailsUtil.toDto(billingdetails);
+            return dto;
+        });
+        
+        log.info("Exiting getAllBillingdetails method");
+        return billingdetailsDtoPage;
     }
   
     /**
@@ -108,20 +109,16 @@ public class BillingdetailsServiceImpl implements BillingdetailsService {
     */
     @Override
     public List<BillingdetailsDto>  searchBillingdetails(Map<String, String> allParams) {
-    log.info("Entering searchBillingdetails method for SQL");
+        log.info("Entering searchBillingdetails method for SQL");
+        
+        Specification<Billingdetails> specification = patientDetailsUtil.constructSearchQuery(allParams, Billingdetails.class);
+        List<Billingdetails> results = billingdetailsRepository.findAll(specification);
+        List<BillingdetailsDto> billingdetailsDtos = results.stream()
+                .map(billingdetails -> patientDetailsUtil.toDto(billingdetails))
+                .collect(Collectors.toList());
 
-    Specification<Billingdetails> specification = constructQuery.constructSearchQuery(allParams, Billingdetails.class);
-    List<Billingdetails> results = billingdetailsRepository.findAll(specification);
-    List<BillingdetailsDto> billingdetailsDtos = results.stream()
-        .map(billingdetails -> {
-            BillingdetailsDto dto = new BillingdetailsDto();
-            BeanUtils.copyProperties(billingdetails, dto);
-            return dto;
-        })
-        .collect(Collectors.toList());
-
-    log.info("Exiting searchBillingdetails method for SQL. Results found: {}", billingdetailsDtos.size());
-    return billingdetailsDtos;
+        log.info("Exiting searchBillingdetails method for SQL. Results found: {}", billingdetailsDtos.size());
+        return billingdetailsDtos;
     }
   
     /**
@@ -132,21 +129,21 @@ public class BillingdetailsServiceImpl implements BillingdetailsService {
     * @throws EntityNotFoundException If no billingdetails with the specified ID is found.
     */
     @Override
-    public BillingdetailsDto  updateBillingdetails(BillingdetailsDto billingdetailsDto) {
-    log.info("Entering updateBillingdetails method for ID: {}", billingdetailsDto.getId());
-    return billingdetailsDao.getBillingdetailsById(billingdetailsDto.getId())
-        .map(existingBillingdetails -> {
-            BeanUtils.copyProperties(billingdetailsDto, existingBillingdetails);
-            Billingdetails updatedBillingdetails = billingdetailsDao.createBillingdetails(existingBillingdetails);
-            BillingdetailsDto responseDto = new BillingdetailsDto();
-            BeanUtils.copyProperties(updatedBillingdetails, responseDto);
-            log.info("Exiting updateBillingdetails method for ID: {}", billingdetailsDto.getId());
-            return responseDto;
-        })
-        .orElseThrow(() -> {
-            log.warn("No billingdetails found for update with ID: {}", billingdetailsDto.getId());
-            return new EntityNotFoundException("Data not found for update with ID: " + billingdetailsDto.getId());
+    public BillingdetailsDto updateBillingdetails(BillingdetailsDto billingdetailsDto) {
+        String id = billingdetailsDto.getId();
+        log.info("Entering updateBillingdetails method for ID: {}", id);
+
+        billingdetailsDao.getBillingdetailsById(id).orElseThrow(() -> {
+            log.warn("No billingdetails found for update with ID: {}", id);
+            return new EntityNotFoundException("Data not found for update with ID: " + id);
         });
+
+        Billingdetails updatedBillingdetails = patientDetailsUtil.toEntity(billingdetailsDto);
+        updatedBillingdetails = billingdetailsDao.createBillingdetails(updatedBillingdetails);
+        BillingdetailsDto responseDto = patientDetailsUtil.toDto(updatedBillingdetails);
+
+        log.info("Exiting updateBillingdetails method for ID: {}", id);
+        return responseDto;
     }
   
     /**
@@ -158,18 +155,17 @@ public class BillingdetailsServiceImpl implements BillingdetailsService {
     */
     @Override
     public String  deleteBillingdetails(String id) {
-    log.info("Entering deleteBillingdetails method for ID: {}", id);
+        log.info("Entering deleteBillingdetails method for ID: {}", id);
 
-    Billingdetails billingdetails = billingdetailsDao.getBillingdetailsById(id)
-         .orElseThrow(() -> {
-             log.warn("No billingdetails found with ID: {}. Deletion failed.", id);
-             return new EntityNotFoundException("No billingdetails found with ID: " + id + ". Unable to delete.");
-         });
+        if (billingdetailsDao.getBillingdetailsById(id).isEmpty()) {
+            log.warn("No billingdetails found with ID: {}. Deletion failed.", id);
+            throw new EntityNotFoundException("No billingdetails found with ID: " + id + ". Unable to delete.");
+        }
 
-    billingdetailsDao.deleteBillingdetails(id);
-    log.info("Successfully deleted Billingdetails with ID: {}", id);
+        billingdetailsDao.deleteBillingdetails(id);
 
-       return "Billingdetails deleted successfully";
+        log.info("Successfully deleted Billingdetails with ID: {}", id);
+        return "Billingdetails deleted successfully";
     }
   
 }
